tests/wpt/web-platform-tests/css/css-scroll-snap/parsing/scroll-margin-block-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-margin-block with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-margin-block">
<meta name="assert" content="scroll-margin-block supports only the grammar '<length>{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("scroll-margin-block-start", "auto");
test_invalid_value("scroll-margin-block-start", "20%");
test_invalid_value("scroll-margin-block-start", "-30%");
test_invalid_value("scroll-margin-block-start", "1px 2px");


test_invalid_value("scroll-margin-block-end", "auto");
test_invalid_value("scroll-margin-block-end", "20%");
test_invalid_value("scroll-margin-block-end", "-30%");
test_invalid_value("scroll-margin-block-end", "1px 2px");


test_invalid_value("scroll-margin-block", "auto");
test_invalid_value("scroll-margin-block", "20%");
test_invalid_value("scroll-margin-block", "-30%");
test_invalid_value("scroll-margin-block", "1px 2px 3px");
</script>
</body>
</html>
